removed unnessacery code

+++ src/Controller/UsersController.php
@@ -5,13 +5,16 @@ namespace App\Controller;
 use App\Entity\Users;
 use App\Form\UsersType;
 use App\Repository\UsersRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
+ *
  * @Route("/users")
+ *
  */
 class UsersController extends AbstractController
 {
@@ -20,6 +23,7 @@ class UsersController extends AbstractController
      */
     public function index(UsersRepository $usersRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MANAGER');
         return $this->render('users/index.html.twig', [
             'users' => $usersRepository->findAll(),
         ]);
@@ -30,14 +34,12 @@ class UsersController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MANAGER');
         $user = new Users();
         $form = $this->createForm(UsersType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $data=$form->getData();
-            var_dump($data);
-            //$user->setRoles([$data['users']['roles']]);
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($user);
             $entityManager->flush();
@@ -56,6 +58,7 @@ class UsersController extends AbstractController
      */
     public function show(Users $user): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MANAGER');
         return $this->render('users/show.html.twig', [
             'user' => $user,
         ]);
@@ -66,6 +69,7 @@ class UsersController extends AbstractController
      */
     public function edit(Request $request, Users $user): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MANAGER');
         $form = $this->createForm(UsersType::class, $user);
         $form->handleRequest($request);
 
@@ -86,6 +90,7 @@ class UsersController extends AbstractController
      */
     public function delete(Request $request, Users $user): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MANAGER');
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
